Add type check paramter in methods and api refactor
