+++ datamegh/api/storage/s3/s3.py
@@ -4,13 +4,13 @@ from boto3.s3.transfer import S3Transfer
 from boto3.exceptions import S3UploadFailedError
 from botocore.exceptions import ClientError
 
-from datamegh.commons.aws import get_aws_client
+from datamegh.api.common.aws import get_aws_client
 from datamegh.util.logging import get_logger
 
 logger = get_logger("storage:s3")
 
 
-def upload(file_path, bucket_name, s3_object_name):
+def upload(file_path: str, bucket_name: str, s3_object_name: str) -> None:
     """
     Method to upload file to s3 bucket
     :param file_path: file path
@@ -34,7 +34,9 @@ def upload(file_path, bucket_name, s3_object_name):
         )
 
 
-def download(file_dir, file_name, bucket_name, s3_object_name):
+def download(
+    file_dir: str, file_name: str, bucket_name: str, s3_object_name: str
+) -> None:
     """
     Method to download file from s3
     :param file_dir: destination file directory eg:/home/destination/
